everything building with trace settings

+++ barretenberg/cpp/src/barretenberg/plonk_honk_shared/arithmetization/arithmetization.hpp
@@ -34,7 +34,7 @@ enum class TraceStructure { NONE, TINY_TEST, SMALL_TEST, CLIENT_IVC_BENCH, E2E_F
 
 struct TraceSettings {
     TraceStructure structure = TraceStructure::NONE;
-    size_t overflow = 0;
+    uint32_t overflow = 0;
 };
 
 /**
